requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
absl-py==0.8.1
addict==2.2.1
albumentations==0.4.5
astor==0.8.0
attrs==19.3.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.28.1
h5py==2.10.0
idna==2.8
imagecorruptions==1.0.0
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.6.0
joblib==0.14.1
jpeg4py==0.1.4
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.32.0
Markdown==3.1.1
matplotlib==3.1.3
mkl
mmcv==0.2.14
networkx==2.4
numba==0.49.0
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.1.0
ordered-set==4.0.1
pandas==1.0.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.19
PyLaTeX==1.3.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.2.1
Shapely==1.7.0
six==1.14.0
sklearn==0.0
tb-nightly==2.2.0a20200205
tensorboard==1.15.0
tensorflow-gpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminaltables==3.1.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.37.0
urllib3==1.25.9
visdom==0.1.8.9
vot-trax==3.0.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.1.0
